RxJava, update source
